Add new build system for the website (#5)
Switches from the ad-hoc description in the README.md file to a Meson
based build. This build provides run targets for deploying and viewing
the v1 and v2 docs. It isn't as good as I want it to be, but I feel it
is a lot easier to use, which is good.

This also opens the door for adding the Javadocs for hse-java to the
site soon.

Signed-off-by: Tristan Partin <tpartin@micron.com>